set memory

+++ 0x07-pointers_arrays_strings/0-memset.c
@@ -10,7 +10,7 @@ char *_memset(char *s, char b, unsigned int n)
 {
 	while (n--)
 	{
-		*s++ = b;
+		*s++ = (unsigned int)b;
 	}
 	return (s);
 }
